clean up interface and tests

+++ src/mcm.jl
@@ -3,7 +3,7 @@ import .MonteCarloMeasurements: Particles
 # From simulation result to particles
 
 vv2m(x) = copy(reduce(hcat, x)')
-function simulate(pf, n, du, Npart)
+function simulate(pf, n, du, Npart::Int)
     sims = map(_->vv2m.(simulate(pf, n, du)), 1:Npart)
     ns = length(sims[1]) # number of time series, usually x,u,y
     ntuple(ns) do s
